Removed some hard-coded attributes
Some of the element attribute values were hard-coded
away from the theme section of the code, so they've
been moved up there.

I expect that the code towards the bottom will be
reduced so it's essentially not repeating each
line of code used to create various css attributes
for the element.