Styles are now pulled dynamically from objects
Instead of applying styled based on what is originally set towards
the top of the file, these are now set and are dynamically applied
given that it can use the node.style.attr api.

Important parts of the data, id, contentEditable etc... are still
set in the same places and are kept away from the styles.